feat(relay): assemble the relay in `Relay::load` instead of `main`
The startup sequence lived in `main.rs`, so an embedder that wants the relay
plus its own workers -- extra routes on the web server, a recorder against
`Cluster::origin`, another listener in its own select! -- had no choice but to
copy `main`'s body. That copy then rots. This crate is consumed as a git
dependency, so a step added here arrives as a library update with no call site
on the other end, and the embedder silently keeps the old behavior: the config
still parses, nothing errors, the feature simply does nothing.

That is not hypothetical. Three omissions of exactly this shape ran on a
production fleet built on this crate: an unbounded group cache while
`--cache-capacity` was set and parsed, a `/nodes` endpoint answering 200 with an
empty node list, and a dropped `--auth-mtls-tier` that would have metered
cert-verified peers on the default tier.

So make the assembly the public surface. `Relay::load` performs the sequence and
returns the pieces; `Relay::run` drives the stock loop; `main` is now a wrapper
over the two. Embedders destructure a `Relay` and drive their own loop, so a
step added to `load` reaches them on their next update, and reshaping the API is
a compile error rather than a silent divergence.

`serve` becomes public for the same reason: an embedder running its own select!
still needs the accept loop, and reimplementing it means re-deriving details
like drawing the `conn` id from the cluster's shared counter.

No behavior change to the standalone binary. The one difference is that the
cluster now gets a clone of the client so `Relay` can expose it, which is how an
embedder reuses the mesh's connection config for its own dials; `Client` is a
handle struct with no `Drop`, so the extra clone is inert.
